projet/tesp/requirements.txt
mysql_connector_repackaged==0.3.1
pip==20.0.2
py_cpuinfo==8.0.0
autopep8==1.6.0
blinker==1.4
browser-cookie3==0.12.1
certifi==2019.11.28
cfgv==3.3.1
chardet==3.0.4
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docopt==0.6.2
emoji==1.2.0
entrypoints==0.3
filelock==3.4.2
Flask==2.0.1
fuse-python==1.0.0
gpg===1.13.1-unknown
GPUtil==1.4.0
httplib2==0.14.0
identify==2.4.9
idna==2.8
importlib-resources==5.4.0
irc==19.0.1
itsdangerous==2.0.1
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.logging==3.1.0
jaraco.stream==3.0.3
jaraco.text==3.7.0
Jinja2==3.0.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lz4==3.1.10
MarkupSafe==2.0.1
millify==0.1.1
more-itertools==8.12.0
mysql-connector==2.2.9
netifaces==0.10.4
nodeenv==1.6.0
oauthlib==3.1.0
olefile==0.46
pbkdf2==1.3
pexpect==4.6.0
Pillow==9.0.0
pipreqs==0.4.11
platformdirs==2.5.0
pre-commit==2.13.0
psutil==5.9.0
py-cpuinfo==8.0.0
pyaes==1.6.1
pycairo==1.16.2
pycodestyle==2.8.0
pycrypto==2.6.1
pycryptodome==3.14.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylibacl==0.5.4
pymacaroons==0.13.0
PyNaCl==1.3.0
PyQt5==5.15.4
PyQt5-sip==12.9.0
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.1
pytz==2021.3
pyxattr==0.6.1
PyYAML==5.3.1
reportlab==3.5.34
requests==2.25.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
tabulate==0.8.9
tempora==5.0.1
toml==0.10.2
tornado==5.1.1
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.8
virtualenv==20.13.1
wadllib==1.3.3
websocket-client==1.0.1
Werkzeug==2.0.3
xkit==0.0.0
yarg==0.1.9
zipp==3.7.0
